Add vim-key navigation to the hexdumper widget
This is a hand-coded widget that does not rely on standard composite
widgets. That was done because constructing the hexdumper widget using
dozens of piles and columns for capturing each hex byte resulted in a
sluggish experience on slow machines like the raspberry pi. So, I need
to manually add vim-key navigation.